+++ packages/@roots/bud-tailwindcss/package.json
@@ -60,6 +60,7 @@
     "pkg:format": "yarn g:pkg:format"
   },
   "devDependencies": {
+    "@roots/bud-api": "workspace:packages/@roots/bud-api",
     "@roots/bud-extensions": "workspace:packages/@roots/bud-extensions",
     "@roots/bud-framework": "workspace:packages/@roots/bud-framework",
     "@roots/bud-postcss": "workspace:packages/@roots/bud-postcss",
@@ -73,7 +74,6 @@
   "peerDependencies": {
     "@roots/bud": "*",
     "@roots/bud-postcss": "*",
-    "@tailwindcss/jit": "^0.1.18",
     "autoprefixer": "^10.2.5",
     "tailwindcss": "^2.1.2"
   }
