+++ packages/@roots/bud-tailwindcss/src/api.ts
@@ -1,26 +1,31 @@
 import {Tailwind} from './interface'
+import {Framework} from '@roots/bud-framework'
 
 const tailwind: Tailwind.Configure = function (
-  config: Omit<Tailwind.Config, null> = null,
+  this: Framework,
   implementation:
     | 'tailwindcss'
     | '@tailwindcss/jit' = 'tailwindcss',
+  config?: Omit<Tailwind.Config, null> | string,
 ) {
-  config = config ?? null
+  const use: [
+    'tailwindcss' | '@tailwindcss/jit',
+    (Omit<Tailwind.Config, null> | string)?,
+  ] = [
+    implementation,
+    config ?? this.path('project', 'tailwind.config.js'),
+  ]
 
+  const plugins = [use, ...Object.values(this.postcss.plugins)]
   if (this.postcss.plugins['postcss-import']) {
     delete this.postcss.plugins['postcss-import']
 
     this.postcss.setPlugins([
       'postcss-import',
-      [implementation, config],
-      ...Object.values(this.postcss.plugins),
-    ])
+      ...plugins,
+    ] as any)
   } else {
-    this.postcss.setPlugins([
-      [implementation, config],
-      ...Object.values(this.postcss.plugins),
-    ])
+    this.postcss.setPlugins([use, ...plugins] as any)
   }
 
   return this
